Check for existence of bc

+++ test/bc/test_bc.sh
@@ -7,6 +7,12 @@ export AFL_I_DONT_CARE_ABOUT_MISSING_CRASHES=1
 user=$(whoami)
 session=/tmp/tmp.${user}.zafl.bc.$$
 
+which bc
+if [ ! $? -eq 0 ]; then
+	echo "SKIP test: bc not on system"
+	exit 0
+fi
+
 cleanup()
 {
 	rm -fr $session
